requirements.txt
django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gevent==21.12.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
pytest==7.1.1
pytest-django==4.5.2
python-dotenv==0.17.1
requests==2.27.1
